bulk commit

+++ myapp/views.py
@@ -1,5 +1,8 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import User
 from django.contrib import messages
 from .models import Room, Topic
@@ -18,19 +21,37 @@ def index(request):
 
 
 def loginPage(request):
+
+    # if user is already logged in then redirect to home. Prevent relogin
+    if request.user.is_authenticated:
+        return redirect("home")
+
     if request.method == "POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
 
         try:
             user = User.objects.get(username=username)
-        except:
+            user = authenticate(request, username=username, password=password)
+
+            if user is not None:
+                login(request, user)  # create session
+                return redirect("home")
+            else:
+                messages.error(request, "Username or password does not match")
+        except User.DoesNotExist:
             messages.error(request, "User does not exist")
 
     context = {}
     return render(request, "myapp/login_register.html", context)
 
 
+def logoutUser(request):
+    logout(request)
+    return redirect("login")
+
+
+@login_required(login_url="login")  # only logged in user can access
 def home(request):
     # for searching on the basis of topic or room or description
     q = request.GET.get("q") if request.GET.get("q") != None else ""
@@ -46,12 +67,14 @@ def home(request):
     return render(request, "myapp/home.html", context)
 
 
+@login_required(login_url="login")
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {"room": room}
     return render(request, "myapp/room.html", context)
 
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -64,9 +87,14 @@ def createRoom(request):
     return render(request, "myapp/room_form.html", context)
 
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    # only host can update
+    if request.user != room.host:
+        return HttpResponse("Only host can update room")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -78,8 +106,12 @@ def updateRoom(request, pk):
     return render(request, "myapp/room_form.html", context)
 
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("Only host can delete room")
     if request.method == "POST":
         room.delete()
         return redirect("home")
